requirements.txt
cryptography==41.0.3
matplotlib==3.7.1
numpy==1.24.3
Pillow==10.1.0
pycryptodome==3.19.0
scapy==2.5.0
Unidecode==1.3.7
ipynb==0.5.1
nbimporter==0.3.4